A dedicated method to use the estimation routines from the VineCopula package

Bivariate copulas are estimated based on BiCopEst and vine copulas through RVineStructureSelect or RVineCopSelect depending on the method argument.

BCfitCopula(copula, data, method = "ml")

Arguments

copula

an object of the desired copula class

data

a matrix holding the U(0,1) distributed data columns

method

for BIVARIATE copulas either "ml" or "itau" for maximum likelihood estimation or inverse tau estimation (for one parameter families) respectively. See BiCopEst for details. In case of a VINE copulas a list with names entries StructureSelect (default: FALSE), indeptest (default: FALSE), familyset (default: 'NA') and indeptest (default: FALSE). See RVineStructureSelect or RVineCopSelect for details.

Value

an object of class fitCopula as in the copula package.

Examples

u <- rCopula(1000, tawnT1Copula(c(3, 0.5)))

fitCopula(tawnT1Copula(), u)
#> Found more than one class "tawnT1Copula" in cache; using the first, from namespace 'VC2copula'
#> Also defined by ‘VineCopula’
#> Call: fitCopula(copula, data = data)
#> Fit based on "mle" and 1000 2-dimensional observations.
#> Copula: tawnT1Copula 
#> param1 param2 
#> 3.2225 0.4934 
#> The maximized loglikelihood is 326