Joe copula models

Wrapper classes representing the bivariate Joe, survival Joe, 90 degree and 270 degree rotated Joe copula families (Joe 1997) from VineCopula::VineCopula-package. Note that package copula-package() provides a class joeBiCopula as well.

Objects from the Classes

Objects can be created by calls of the form new("joeBiCopula", ...), new("surJoeBiCopula", ...), new("r90JoeBiCopula", ...) and new("r270JoeBiCopula", ...) or by the functions joeBiCopula(), surJoeBiCopula(), r90JoeBiCopula() and r270JoeBiCopula().

References

Joe, H., (1997). Multivariate Models and Dependence Concepts. Monogra. Stat. Appl. Probab. 73, London: Chapman and Hall.

See also

Examples

showClass("surJoeBiCopula")
#> Found more than one class "surJoeBiCopula" in cache; using the first, from namespace 'VC2copula'
#> Also defined by ‘VineCopula’
#> Found more than one class "surJoeBiCopula" in cache; using the first, from namespace 'VC2copula'
#> Also defined by ‘VineCopula’
#> Class "surJoeBiCopula" [package "VC2copula"]
#> 
#> Slots:
#>                                                                        
#> Name:        family   parameters  param.names param.lowbnd  param.upbnd
#> Class:      numeric      numeric    character      numeric      numeric
#>                                 
#> Name:      fullname    dimension
#> Class:    character      integer
#> 
#> Extends: 
#> Class "copula", directly
#> Class "dimCopula", by class "copula", distance 2
#> Class "parCopula", by class "copula", distance 2
#> Class "Copula", by class "copula", distance 3