Tawn copula models (type 2)

S4-class representation of the Tawn Copula family of type 2 and rotated versions there of.

Objects from the Class

Objects can be created by calls of the form new("tawnT2Copula", ...), or through the explicit constructors tawnT2Copula(), surTawnT2Copula(), r90TawnT2Copula() and r270TawnT2Copula() respectively.

See also

Examples

showClass("tawnT2Copula")

#> Found more than one class "tawnT2Copula" in cache; using the first, from namespace 'VC2copula'

#> Also defined by ‘VineCopula’

#> Found more than one class "tawnT2Copula" in cache; using the first, from namespace 'VC2copula'

#> Also defined by ‘VineCopula’

#> Class "tawnT2Copula" [package "VC2copula"]
#> 
#> Slots:
#>                                                                        
#> Name:        family   parameters  param.names param.lowbnd  param.upbnd
#> Class:      numeric      numeric    character      numeric      numeric
#>                                 
#> Name:      fullname    dimension
#> Class:    character      integer
#> 
#> Extends: 
#> Class "copula", directly
#> Class "dimCopula", by class "copula", distance 2
#> Class "parCopula", by class "copula", distance 2
#> Class "Copula", by class "copula", distance 3